renderservice/requirements.txt
flask==0.12.1
gevent==1.2.1
gunicorn==19.7.1
google-cloud==0.24.0
google-api-python-client==1.6.2
daemons==1.3.0
jinja2==2.9.6
Pillow==4.1.1
urllib3[secure]==1.21.1
httplib2shim==0.0.2
yattag==1.9.0

# Testing
coverage==4.4.1
flake8==3.3.0
pydocstyle==2.0.0
